+++ src/controls/polyControl.spec.ts
@@ -19,10 +19,10 @@ describe('polyControl', () => {
     canvas
       .getSelectionElement()
       .dispatchEvent(new MouseEvent('mousedown', { clientX: 50, clientY: 50 }));
-    canvas._setupCurrentTransform(
+    canvas.setupCurrentTransform(
       new MouseEvent('mousedown', { clientX: 50, clientY: 50 }),
       poly,
-      true
+      poly.findControl(new Point(50, 50))
     );
     document.dispatchEvent(
       new MouseEvent('mousemove', { clientX: 55, clientY: 55 })